Switch over from Double to BigDecimal with a basic implementation of pow and sqrt for now. Also check for null input when equals is tapped.
TODO: Change tests to BigDecimal.

+++ app/src/main/java/com/example/bettercalc/parsing/Expression.kt
@@ -1,9 +1,9 @@
 package com.example.bettercalc.parsing
 
-import com.example.bettercalc.parsing.TokenType
+import java.math.BigDecimal
 
 abstract class Expression {
 //    abstract override fun toString(): String
-    abstract fun calculate(): Double
+    abstract fun calculate(): BigDecimal
     abstract fun fetchOperator(): TokenType?
 }
